interaction between formatting and elastic tabstops
This commit adds a __format action__ to the editor.
The formatter is _google java formatter_ and as such can only handle
Java code.
It was chosen because of its maturity and my familiarity with Java.

The format action simply formats the text buffer in the most naiv way
and marks the buffe as modified.
I expect to have to tinker around wth the ´spacesToTabs´ and
´tabsToSpaces´ functions in a future commit.